requirements.txt
Django==3.2.7

django-environ==0.6.0
django-cors-headers==3.8.0
django-redis==5.0.0
django-filter==2.4.0
django-model-utils==4.1.1

djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0

djoser==2.1.0
drf-yasg==1.20.0
drf-access-policy==1.1.0
drf-writable-nested==0.6.3

gunicorn==20.1.0
uvicorn==0.15.0

psycopg2-binary==2.9.1
pydantic==1.8.2
orjson==3.6.3
